Added error logging

+++ src/utilities/VirtualFileSystem.cs
@@ -4,6 +4,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Runtime.InteropServices;
+using System.Text;
+using BG3ModdingUtil;
 
 namespace SSA.VirtualFileSystem
 {
@@ -14,6 +16,38 @@ namespace SSA.VirtualFileSystem
         //https://learn.microsoft.com/en-us/dotnet/api/system.io.file.createsymboliclink?view=net-7.0
         //https://stackoverflow.com/questions/3387690/how-to-create-a-hardlink-in-c
         //https://github.com/usdAG/SharpLink   
+
+        public void GetException(Exception e, string info, DirectoryInfo location){
+            StringBuilder sb = new();
+            sb.AppendLine(info);
+            string str = string.Format("Exception Message: {0}", e.Message);
+            sb.AppendLine(str);
+            str = string.Format("Exception Source: {0}", e.Source);
+            sb.AppendLine(str);
+            str = string.Format("Inner Exception: {0}", e.InnerException);
+            sb.AppendLine(str);
+            str = string.Format("Base Exception: {0}", e.GetBaseException().ToString());
+            sb.AppendLine(str);
+            str = string.Format("Exception Data: {0}", e.Data);
+            sb.AppendLine(str);
+            str = string.Format("Exception Stack Trace: {0}", e.StackTrace);
+            sb.AppendLine(str);
+            List<string> dirs = Directory.GetDirectories(location.FullName).ToList();
+            List<string> files = Directory.GetFiles(location.FullName).ToList();
+            StringBuilder filesandfolders = new();
+
+            sb.AppendLine("Items in Location:");            
+            foreach (string folder in dirs){
+                DirectoryInfo dirInfo = new(folder);
+                filesandfolders.AppendLine(string.Format("       Folder: {0}\n       Symlink: {1}", dirInfo.Name, dirInfo.Attributes.HasFlag(FileAttributes.ReparsePoint).ToString()));
+            }          
+            foreach (string file in files){
+                FileInfo fileInfo = new(file);
+                filesandfolders.AppendLine(string.Format("       File: {0}\n       Symlink: {1}", fileInfo.Name, fileInfo.Attributes.HasFlag(FileAttributes.ReparsePoint).ToString()));
+            }
+            sb.AppendLine(filesandfolders.ToString());
+            Globals.MakeExceptionReport(sb);
+        }
    
         
         public void MakeSymbolicLink(string Original, string Destination){
@@ -23,15 +57,18 @@ namespace SSA.VirtualFileSystem
                 try { 
                     File.Move(Destination, string.Format("{0}.disabled", Destination)); 
                 } catch (Exception e) {
-                    System.Windows.Forms.MessageBox.Show(string.Format("Caught exception disabling duplicate file: {0}", e.Message), fileInfo.Name);
+                    string exceptionstring = string.Format("Caught exception disabling duplicate file: {0}\nException: {1}", fileInfo.Name, e.Message);
+                    System.Windows.Forms.MessageBox.Show(exceptionstring);
+                    GetException(e, exceptionstring, fileInfo.Directory);
                 }
             }
             try {
                 File.CreateSymbolicLink(Destination, Original);
             } catch (Exception e) {
-                System.Windows.Forms.MessageBox.Show(string.Format("Caught exception making symbolic link: {0}", e.Message), fileInfo.Name);
-            }
-            
+                string exceptionstring = string.Format("Caught exception making symbolic link: {0}\nException: {1}", fileInfo.Name, e.Message);
+                System.Windows.Forms.MessageBox.Show(exceptionstring);
+                GetException(e, exceptionstring, fileInfo.Directory);
+            }            
         }
 
         public void MakeSymbolicLink(string Original, string Destination, string AsName){
@@ -41,13 +78,17 @@ namespace SSA.VirtualFileSystem
                 try {
                     File.Move(Destination, string.Format("{0}.disabled", Destination));
                 } catch (Exception e) {
-                    System.Windows.Forms.MessageBox.Show(string.Format("Caught exception disabling duplicate file: {0}", e.Message), fileInfo.Name);
+                    string exceptionstring = string.Format("Caught exception disabling duplicate file: {0}\nException: {1}", fileInfo.Name, e.Message);
+                    System.Windows.Forms.MessageBox.Show(exceptionstring);
+                    GetException(e, exceptionstring, fileInfo.Directory);
                 }
             }
             try {
                 File.CreateSymbolicLink(Destination, Original);
             } catch (Exception e) {
-                System.Windows.Forms.MessageBox.Show(string.Format("Caught exception making symbolic link with custom name: {0}", e.Message), fileInfo.Name);
+                string exceptionstring = string.Format("Caught exception making symbolic link: {0}\nException: {1}", fileInfo.Name, e.Message);
+                System.Windows.Forms.MessageBox.Show(exceptionstring);
+                GetException(e, exceptionstring, fileInfo.Directory);
             }
             
         }
@@ -57,7 +98,9 @@ namespace SSA.VirtualFileSystem
                 try { 
                     File.Delete(Item); 
                 } catch (Exception e) {
-                    System.Windows.Forms.MessageBox.Show(string.Format("Caught exception deleting symbolic link: {0}", e.Message), Item);
+                    string exceptionstring = string.Format("Caught exception deleting symbolic link: {0}\nException: {1}", Item, e.Message);
+                    System.Windows.Forms.MessageBox.Show(exceptionstring);
+                    GetException(e, exceptionstring, new FileInfo(Item).Directory);
                 }
             }            
             if (File.Exists(string.Format("{0}.disabled", Item))){
@@ -66,7 +109,9 @@ namespace SSA.VirtualFileSystem
                 try {
                     File.Move(og, ren);
                 } catch (Exception e) {
-                    System.Windows.Forms.MessageBox.Show(string.Format("Caught exception renaming disabled file: {0}", e.Message), Item);
+                    string exceptionstring = string.Format("Caught exception renaming disabled file: {0}\nException: {1}", Item, e.Message);
+                    System.Windows.Forms.MessageBox.Show(exceptionstring);
+                    GetException(e, exceptionstring, new FileInfo(Item).Directory);
                 }
             }
         }
@@ -78,13 +123,17 @@ namespace SSA.VirtualFileSystem
                 try { 
                     Directory.Move(Destination, string.Format("{0}--DISABLED", Destination)); 
                 } catch (Exception e) {
-                    System.Windows.Forms.MessageBox.Show(string.Format("Caught exception disabling duplicate folder: {0}", e.Message), directoryInfo.Name);
+                    string exceptionstring = string.Format("Caught exception disabling duplicate folder: {0}\nException: {1}", directoryInfo.Name, e.Message);
+                    System.Windows.Forms.MessageBox.Show(exceptionstring);
+                    GetException(e, exceptionstring, directoryInfo);
                 }
             }
             try {
                 Directory.CreateSymbolicLink(Destination, Original);
-            } catch (Exception e) {
-                System.Windows.Forms.MessageBox.Show(string.Format("Caught exception making junction: {0}", e.Message), directoryInfo.Name);
+            } catch (Exception e) {         
+                string exceptionstring = string.Format("Caught exception making junction: {0}\nException: {1}", directoryInfo.Name, e.Message);
+                System.Windows.Forms.MessageBox.Show(exceptionstring);
+                GetException(e, exceptionstring, directoryInfo);
             }
         }
 
@@ -93,16 +142,24 @@ namespace SSA.VirtualFileSystem
                 try {
                     Directory.Delete(Item);
                 } catch (Exception e) {
-                    System.Windows.Forms.MessageBox.Show(string.Format("Caught exception removing junction: {0}", e.Message), Item);
+                    string exceptionstring = string.Format("Caught exception removing junction: {0}\nException: {1}", Item, e.Message);
+                    System.Windows.Forms.MessageBox.Show(exceptionstring);
+                    GetException(e, exceptionstring, new FileInfo(Item).Directory);
                 }
             }            
             if (Directory.Exists(string.Format("{0}--DISABLED", Item))){
                 string og = string.Format("{0}--DISABLED", Item);
                 string ren = og.Replace("--DISABLED", "");
                 try {
-                    File.Move(og, ren);
+                    if(new DirectoryInfo(Item).Attributes.HasFlag(FileAttributes.ReparsePoint)){
+                        Directory.Move(og, ren);
+                    } else {
+                        File.Move(og, ren);
+                    }                    
                 } catch (Exception e) {
-                    System.Windows.Forms.MessageBox.Show(string.Format("Caught exception renaming disabled folder: {0}", e.Message), Item);
+                    string exceptionstring = string.Format("Caught exception renaming disabled folder: {0}\nException: {1}", Item, e.Message);
+                    System.Windows.Forms.MessageBox.Show(exceptionstring);
+                    GetException(e, exceptionstring, new FileInfo(Item).Directory);
                 }
             }
         }
